move tooltip to its own functional component
