+++ ui/components/app/edit-gas-fee-popover/edit-gas-item/edit-gas-item.js
@@ -26,6 +26,7 @@ import {
   COLORS,
   FONT_WEIGHT,
 } from '../../../../helpers/constants/design-system';
+import EditGasToolTip from '../edit-gas-tooltip';
 import { useCustomTimeEstimate } from './useCustomTimeEstimate';
 
 const EditGasItem = ({ priorityLevel }) => {
@@ -54,13 +55,9 @@ const EditGasItem = ({ priorityLevel }) => {
     priorityLevel === PRIORITY_LEVELS.DAPP_SUGGESTED &&
     dappSuggestedGasFees
   ) {
-    maxFeePerGas = hexWEIToDecGWEI(
-      dappSuggestedGasFees.maxFeePerGas ?? dappSuggestedGasFees.gasPrice,
-    );
+    maxFeePerGas = hexWEIToDecGWEI(dappSuggestedGasFees.maxFeePerGas);
     maxPriorityFeePerGas = hexWEIToDecGWEI(
-      dappSuggestedGasFees.maxPriorityFeePerGas ??
-        // TODO should this default back to something else? Will definitely be too high in this case
-        dappSuggestedGasFees.gasPrice,
+      dappSuggestedGasFees.maxPriorityFeePerGas,
     );
   } else if (priorityLevel === PRIORITY_LEVELS.CUSTOM) {
     if (estimateUsed === PRIORITY_LEVELS.CUSTOM) {
@@ -102,41 +99,6 @@ const EditGasItem = ({ priorityLevel }) => {
     } else {
       updateTransactionUsingGasFeeEstimates(priorityLevel);
       closeModal('editGasFee');
-    }
-  };
-
-  const toolTipMessage = () => {
-    switch (priorityLevel) {
-      case PRIORITY_LEVELS.LOW:
-        return t('lowGasSettingToolTipMessage', [
-          <span key={priorityLevel}>
-            <b>{t('low')}</b>
-          </span>,
-        ]);
-      case PRIORITY_LEVELS.MEDIUM:
-        return t('mediumGasSettingToolTipMessage', [
-          <span key={priorityLevel}>
-            <b>{t('medium')}</b>
-          </span>,
-        ]);
-      case PRIORITY_LEVELS.HIGH:
-        return t('highGasSettingToolTipMessage', [
-          <span key={priorityLevel}>
-            <b>{t('high')}</b>
-          </span>,
-        ]);
-      case PRIORITY_LEVELS.CUSTOM:
-        return t('customGasSettingToolTipMessage', [
-          <span key={priorityLevel}>
-            <b>{t('custom')}</b>
-          </span>,
-        ]);
-      case PRIORITY_LEVELS.DAPP_SUGGESTED:
-        return t('dappSuggestedGasSettingToolTipMessage', [
-          <span key={origin}>{origin}</span>,
-        ]);
-      default:
-        return '';
     }
   };
 
@@ -191,88 +153,16 @@ const EditGasItem = ({ priorityLevel }) => {
         )}
       </span>
       <span className="edit-gas-item__tooltip">
-        <InfoTooltip
-          contentText={
-            <div className="edit-gas-item__tooltip__container">
-              {priorityLevel !== PRIORITY_LEVELS.CUSTOM &&
-              priorityLevel !== PRIORITY_LEVELS.DAPP_SUGGESTED ? (
-                <img
-                  alt=""
-                  width={130}
-                  src={`./images/curve-${priorityLevel}.svg`}
-                />
-              ) : null}
-              {priorityLevel === PRIORITY_LEVELS.HIGH ? (
-                <div className="edit-gas-item__tooltip__container__dialog">
-                  <Typography fontSize="12px" color={COLORS.WHITE}>
-                    {t('highGasSettingToolTipDialog')}
-                  </Typography>
-                </div>
-              ) : null}
-              <div className="edit-gas-item__tooltip__container__message">
-                <Typography fontSize="12px">{toolTipMessage()}</Typography>
-              </div>
-              {priorityLevel === PRIORITY_LEVELS.CUSTOM ? null : (
-                <div className="edit-gas-item__tooltip__container__values">
-                  <div>
-                    <Typography
-                      fontSize="12px"
-                      fontWeight={FONT_WEIGHT.BOLD}
-                      margin={0}
-                      className="edit-gas-item__tooltip__container__label"
-                    >
-                      {t('maxBaseFee')}
-                    </Typography>
-                    <Typography
-                      fontSize="12px"
-                      color={COLORS.NEUTRAL_GREY}
-                      margin={0}
-                      className="edit-gas-item__tooltip__container__value"
-                    >
-                      {maxFeePerGas}
-                    </Typography>
-                  </div>
-                  <div>
-                    <Typography
-                      fontSize="12px"
-                      fontWeight={FONT_WEIGHT.BOLD}
-                      margin={0}
-                      className="edit-gas-item__tooltip__container__label"
-                    >
-                      {t('priorityFee')}
-                    </Typography>
-                    <Typography
-                      fontSize="12px"
-                      color={COLORS.NEUTRAL_GREY}
-                      margin={0}
-                      className="edit-gas-item__tooltip__container__value"
-                    >
-                      {maxPriorityFeePerGas ?? maxPriorityFeePerGasValue}
-                    </Typography>
-                  </div>
-                  <div>
-                    <Typography
-                      fontSize="12px"
-                      fontWeight={FONT_WEIGHT.BOLD}
-                      margin={0}
-                      className="edit-gas-item__tooltip__container__label"
-                    >
-                      {t('gasLimit')}
-                    </Typography>
-                    <Typography
-                      fontSize="12px"
-                      color={COLORS.NEUTRAL_GREY}
-                      margin={0}
-                      className="edit-gas-item__tooltip__container__value"
-                    >
-                      {gasLimit}
-                    </Typography>
-                  </div>
-                </div>
-              )}
-            </div>
-          }
-          position="top"
+        <EditGasToolTip
+          priorityLevel={priorityLevel}
+          gasContext={{
+            gasLimit,
+            maxFeePerGas,
+            maxPriorityFeePerGas,
+            maxFeePerGasValue,
+            maxPriorityFeePerGasValue,
+            origin,
+          }}
         />
       </span>
     </button>
